Move ZLL 5000k bulb to cloud
This move the "ZLL White Color Temperature Bulb 5000k" DTH to execute in
the cloud until 23.  This is because in the current local implementation
there isn't a read after a delay causing us to be out of sync
temporarily.

This addresses: https://smartthings.atlassian.net/browse/ICP-5419
This addresses: https://smartthings.atlassian.net/browse/ICP-5416